// src/types/Asistente.ts
export interface Asistente {
  id?: string;
  rut: string;
  nombre: string;
  apellido: string;
  mail: string;
  celular: string;
  fnac: string;
  especialidad: string;
  region: string;
  comuna: string;
  browserId: string;
  checkboxterminos: string;
  createdAt?: string;
  updatedAt?: string;
} 